backend/requirements.txt
flask
flask-cors
pymysql
bcrypt
requests
urllib3
dbutils
google-genai
